Trait tract_core::framework::Framework

pub trait Framework<ProtoModel, Model> where
    ProtoModel: Debug, {
    fn proto_model_for_read(
        &self, 
        reader: &mut dyn Read
    ) -> TractResult<ProtoModel>;

    fn model_for_proto_model(&self, proto: &ProtoModel) -> TractResult<Model>;

    fn proto_model_for_path(
        &self, 
        p: impl AsRef<Path>
    ) -> TractResult<ProtoModel> { ... }

    fn model_for_read(&self, r: &mut dyn Read) -> TractResult<Model> { ... }

    fn model_for_path(&self, p: impl AsRef<Path>) -> TractResult<Model> { ... }
}

A Framework that translate its model to tract core model.

The ProtoModel is the parsed representation of the imported model. It does not have to be Protobuf based.

Required methods

fn proto_model_for_read(&self, reader: &mut dyn Read) -> TractResult<ProtoModel>

Parse a proto model from a reader.

fn model_for_proto_model(&self, proto: &ProtoModel) -> TractResult<Model>

Translate a proto model into a model.

Provided methods

fn proto_model_for_path(&self, p: impl AsRef<Path>) -> TractResult<ProtoModel>

Read a proto model from a filename.

fn model_for_read(&self, r: &mut dyn Read) -> TractResult<Model>

Read a model from a reader

fn model_for_path(&self, p: impl AsRef<Path>) -> TractResult<Model>

Build a model from a filename.