Struct tract_core::model::ModelPatch

pub struct ModelPatch<F, O> where
    F: Fact + Clone + 'static + Hash,
    O: Display + Debug + AsRef<dyn Op> + AsMut<dyn Op> + Clone + 'static + Hash,  {
    pub context: Vec<String>,
    pub dont_apply_twice: Option<String>,
    pub model: Graph<F, O>,
    pub inputs: HashMap<usize, usize>,
    pub incoming: HashMap<OutletId, OutletId>,
    pub shunt_outlet_by: HashMap<OutletId, OutletId>,
    pub obliterate: Vec<usize>,
}

A change to apply to a model.

Actually structured around a model that represent the new nodes to be inserted, plus information about how to connect these new nodes to the pre-existing graph.

Fields
context: Vec<String>

patch label for auditing and debugging

dont_apply_twice: Option<String>

optimizer will ignore this patch in node to node loop if it was already encountered

model: Graph<F, O>

the model-like ‘patch’ of nodes to add to the model

inputs: HashMap<usize, usize>

map of replaced inputs (patch node id to model node id)

incoming: HashMap<OutletId, OutletId>

map of patch inputs to model wires

shunt_outlet_by: HashMap<OutletId, OutletId>

map of old wires to be replaced by new wires

obliterate: Vec<usize>

operations to discard from the model

Implementations

impl<F, O> ModelPatch<F, O> where
F: Fact + Clone + 'static + Hash,
O: Display + Debug + AsRef<dyn Op> + AsMut<dyn Op> + Clone + 'static + Hash,
Graph<F, O>: SpecialOps<F, O>,

pub fn new(s: impl Into<String>) -> Self

pub fn push_context(&mut self, s: impl Into<String>)

pub fn with_context(self, s: impl Into<String>) -> Self

pub fn is_empty(&self) -> bool

pub fn tap_model(
&mut self,
model: &Graph<F, O>,
outlet: OutletId
) -> TractResult<OutletId>

Draw a tap from a preexisting node.

returns an OutletId usable in the little “patch” model

pub unsafe fn shunt_outside_unchecked(
&mut self,
outlet: OutletId,
by: OutletId
) -> TractResult<()>

pub fn shunt_outside(
&mut self,
model: &Graph<F, O>,
outlet: OutletId,
by: OutletId
) -> TractResult<()>

Replace an Outlet in the target model by one from the patch.

pub fn obliterate(&mut self, node: usize) -> TractResult<()>

pub fn replace_single_op<IO: Into<O>>(
patched_model: &Graph<F, O>,
node: &Node<F, O>,
inputs: &[OutletId],
new_op: IO
) -> TractResult<ModelPatch<F, O>>

Convenience method creating a patch that replace a single operation.

pub fn fuse_with_next<IO: Into<O>>(
patched_model: &Graph<F, O>,
node: &Node<F, O>,
new_op: IO
) -> TractResult<ModelPatch<F, O>>

Convenience method creating a patch that replace a single operation.

pub fn shunt_one_op(
patched_model: &Graph<F, O>,
node: &Node<F, O>
) -> TractResult<ModelPatch<F, O>>

Convenience method creating a patch that shunt the given node.

pub fn single_unary_op<IO: Into<O>>(
patched_model: &Graph<F, O>,
node: &Node<F, O>,
new_op: IO
) -> TractResult<ModelPatch<F, O>>

Convenience method creating a patch that replace a single unary operation.

pub fn intercept<IO: Into<O>>(
patched_model: &Graph<F, O>,
outlet: OutletId,
name: impl Into<String>,
new_op: IO,
fact: F
) -> TractResult<ModelPatch<F, O>>

Convenience method creating a patch that insert an unary op on an outlet.

pub fn apply(self, target: &mut Graph<F, O>) -> TractResult<()>

Apply all changes in the patch to the target model.