Trait tract_core::ops::binary::BinMiniOp

pub trait BinMiniOp: Debug + DynClone + Send + Sync + 'static + Downcast + DynHash {

    fn name(&self) -> &'static str;

    fn operating_datum_type(
        &self, 
        a: DatumType, 
        b: DatumType
    ) -> TractResult<DatumType>;

    fn result_datum_type(
        &self, 
        a: DatumType, 
        b: DatumType
    ) -> TractResult<DatumType>;

    fn eval_unicast_in_place(
        &self, 
        a: &Tensor, 
        b: &mut Tensor
    ) -> TractResult<()>;

    fn eval_uniform_in_place(
        &self, 
        a: &Tensor, 
        b: &mut Tensor
    ) -> TractResult<()>;

    fn eval_out_of_place(
        &self, 
        c: &mut Tensor, 
        a: &Tensor, 
        b: &Tensor
    ) -> TractResult<()>;

    fn validation(&self) -> Validation { ... }

    fn generic_eval(
        &self, 
        a: Arc<Tensor>, 
        b: Arc<Tensor>
    ) -> TractResult<Tensor> { ... }

    fn eval(&self, a: Arc<Tensor>, b: Arc<Tensor>) -> TractResult<Tensor> { ... }

    fn unary_with_b_const(&self, b: &Arc<Tensor>) -> Option<UnaryOp> { ... }

    fn declutter_bin(
        &self, 
        model: &TypedModel, 
        node: &TypedNode
    ) -> TractResult<Option<TypedModelPatch>> { ... }

    fn declutter_unary(
        &self, 
        model: &TypedModel, 
        node: &TypedNode, 
        a: &Arc<Tensor>
    ) -> TractResult<Option<TypedModelPatch>> { ... }

    fn cost_per_element(&self, dt: DatumType) -> TVec<(Cost, usize)> { ... }

}

Required methods

fn name(&self) -> &'static str

fn operating_datum_type(
&self,
a: DatumType,
b: DatumType
) -> TractResult<DatumType>

fn result_datum_type(
&self,
a: DatumType,
b: DatumType
) -> TractResult<DatumType>

fn eval_unicast_in_place(&self, a: &Tensor, b: &mut Tensor) -> TractResult<()>

fn eval_uniform_in_place(&self, a: &Tensor, b: &mut Tensor) -> TractResult<()>

fn eval_out_of_place(
&self,
c: &mut Tensor,
a: &Tensor,
b: &Tensor
) -> TractResult<()>

Provided methods

fn validation(&self) -> Validation

fn generic_eval(&self, a: Arc<Tensor>, b: Arc<Tensor>) -> TractResult<Tensor>

fn eval(&self, a: Arc<Tensor>, b: Arc<Tensor>) -> TractResult<Tensor>

fn unary_with_b_const(&self, b: &Arc<Tensor>) -> Option<UnaryOp>

fn declutter_bin(
&self,
model: &TypedModel,
node: &TypedNode
) -> TractResult<Option<TypedModelPatch>>

fn declutter_unary(
&self,
model: &TypedModel,
node: &TypedNode,
a: &Arc<Tensor>
) -> TractResult<Option<TypedModelPatch>>

fn cost_per_element(&self, dt: DatumType) -> TVec<(Cost, usize)>

Implementations

impl dyn BinMiniOp

pub fn is<__T: BinMiniOp>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: BinMiniOp>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: BinMiniOp>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: BinMiniOp>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: BinMiniOp>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.