Trait tract_core::ops::Op

pub trait Op: Debug + DynClone + Send + Sync + 'static + Downcast + EvalOp + DynHash {
    fn op_families(&self) -> &'static [&'static str];

    fn name(&self) -> Cow<'_, str>;

    fn as_typed(&self) -> Option<&dyn TypedOp>;

    fn validation(&self) -> Validation { ... }

    fn same_as(&self, _other: &dyn Op) -> bool { ... }

    fn info(&self) -> TractResult<Vec<String>> { ... }
}

A base operation

Required methods

fn op_families(&self) -> &'static [&'static str]

Vector of short strings defining what families the op belongs too. tract-core defines “core”, “mir”, “lir”.

fn name(&self) -> Cow<'_, str>

fn as_typed(&self) -> Option<&dyn TypedOp>

Provided methods

fn validation(&self) -> Validation

The kind of accuracy check that should be performed on operation when testing them.

fn same_as(&self, _other: &dyn Op) -> bool

Compare two ops.

fn info(&self) -> TractResult<Vec<String>>

Short (one-line) strings giving hints on internal implementation or important configuration details to be displayed in dumps.

Implementations

impl dyn Op

pub fn is<__T: Op>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Op>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: Op>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: Op>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: Op>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.